+++ CPP04/ex03/include/ICharacter.hpp
@@ -11,8 +11,8 @@ public:
 	virtual ~ICharacter() {}
 	virtual std::string const &getName() const = 0;
 	virtual void equip(AMateria *m) = 0;
-	virtual void unequip(size_t idx) = 0;
-	virtual void use(size_t idx, ICharacter &target) = 0;
+	virtual void unequip(int idx) = 0;
+	virtual void use(int idx, ICharacter &target) = 0;
 
 	virtual void lookAtFloor(void) = 0;
 	virtual void clearFloor(void) = 0;
